Rename df.make -> withSchema - and - df.fromComposable -> applySchema
