+++ examples/remix/app/business/gpd.ts
@@ -6,7 +6,7 @@ const cookie = createCookie('gpd', {
   maxAge: 60, // One minute, but should probably be longer
 })
 
-const getGPDInfo = df.make(
+const getGPDInfo = withSchema(
   z.any(),
   // The "environment" knows there can be cookie information in the Request
   z.object({ agreed: z.boolean().optional() }),
@@ -14,7 +14,7 @@ const getGPDInfo = df.make(
   return { agreed }
 })
 
-const agreeToGPD = df.make(
+const agreeToGPD = withSchema(
   // Agreeing to the GPD is user input
   z.object({ agree: z.preprocess((v) => v === 'true', z.boolean()) }),
 )(async ({ agree }) => ({ agreed: agree }))
