+++ src/df/types.ts
@@ -51,36 +51,7 @@ type UnpackDFObject<Obj extends Record<string, Composable>> =
   | { [K in keyof Obj]: UnpackData<Obj[K]> }
   | never
 
-/**
- * A parsing error when validating the input or environment schemas.
- * This will be transformed into an `InputError` before being returned from the domain function.
- * It is usually not visible to the end user unless one wants to write an adapter for a schema validator.
- */
-type ParserIssue = { path: PropertyKey[]; message: string }
-
-/**
- * The result of input or environment validation.
- * See the type `Result` for the return values of domain functions.
- * It is usually not visible to the end user unless one wants to write an adapter for a schema validator.
- */
-type ParserResult<T> =
-  | {
-      success: true
-      data: T
-    }
-  | { success: false; error: { issues: ParserIssue[] } }
-
-/**
- * The object used to validate either input or environment when creating domain functions.
- */
-type ParserSchema<T extends unknown = unknown> = {
-  safeParseAsync: (a: unknown) => Promise<ParserResult<T>>
-}
-
 export type {
-  ParserIssue,
-  ParserResult,
-  ParserSchema,
   UnpackAll,
   UnpackData,
   UnpackDFObject,
