+++ src/tests/first.test.ts
@@ -1,12 +1,12 @@
 import { assertEquals, describe, it, z } from '../test-prelude.ts'
-import { df, first, failure, InputError, success } from '../index.ts'
+import { withSchema, first, failure, InputError, success } from '../index.ts'
 import type { Composable } from '../index.ts'
 
 describe('first', () => {
   it('should return the result of the first successful domain function', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => String(id))
-    const b = df.make(z.object({ id: z.number() }))(({ id }) => id + 1)
-    const c = df.make(z.object({ id: z.number() }))(({ id }) => Boolean(id))
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => String(id))
+    const b = withSchema(z.object({ id: z.number() }))(({ id }) => id + 1)
+    const c = withSchema(z.object({ id: z.number() }))(({ id }) => Boolean(id))
     const d = first(a, b, c)
     type _R = Expect<
       Equal<
@@ -22,10 +22,10 @@ describe('first', () => {
   })
 
   it('should return a successful result even if one of the domain functions fails', async () => {
-    const a = df.make(
+    const a = withSchema(
       z.object({ n: z.number(), operation: z.literal('increment') }),
     )(({ n }) => n + 1)
-    const b = df.make(
+    const b = withSchema(
       z.object({ n: z.number(), operation: z.literal('decrement') }),
     )(({ n }) => n - 1)
 
@@ -41,8 +41,8 @@ describe('first', () => {
   })
 
   it('should return error when all of the domain functions fails', async () => {
-    const a = df.make(z.object({ id: z.string() }))(({ id }) => id)
-    const b = df.make(z.object({ id: z.number() }))(() => {
+    const a = withSchema(z.object({ id: z.string() }))(({ id }) => id)
+    const b = withSchema(z.object({ id: z.number() }))(() => {
       throw 'Error'
     })
