+++ src/tests/merge.test.ts
@@ -5,15 +5,15 @@ import {
   it,
   z,
 } from '../test-prelude.ts'
-import { merge, df, failure, InputError, success } from '../index.ts'
+import { merge, withSchema, failure, InputError, success } from '../index.ts'
 import type { Composable } from '../index.ts'
 
 describe('merge', () => {
   it('should combine two domain functions results into one object', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       resultA: id + 1,
     }))
-    const b = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const b = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       resultB: id - 1,
     }))
 
@@ -34,15 +34,15 @@ describe('merge', () => {
   })
 
   it('should combine many domain functions into one', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       resultA: String(id),
       resultB: String(id),
       resultC: String(id),
     }))
-    const b = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const b = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       resultB: id + 1,
     }))
-    const c = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const c = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       resultC: Boolean(id),
     }))
     const d = merge(a, b, c)
@@ -67,10 +67,10 @@ describe('merge', () => {
   })
 
   it('should return error when one of the domain functions has input errors', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       id,
     }))
-    const b = df.make(z.object({ id: z.string() }))(({ id }) => ({
+    const b = withSchema(z.object({ id: z.string() }))(({ id }) => ({
       id,
     }))
 
@@ -98,10 +98,10 @@ describe('merge', () => {
   })
 
   it('should return error when one of the domain functions fails', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => ({
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       id,
     }))
-    const b = df.make(z.object({ id: z.number() }))(() => {
+    const b = withSchema(z.object({ id: z.number() }))(() => {
       throw 'Error'
     })
 
@@ -118,10 +118,10 @@ describe('merge', () => {
   })
 
   it('should combine the inputError messages of both functions', async () => {
-    const a = df.make(z.object({ id: z.string() }))(({ id }) => ({
+    const a = withSchema(z.object({ id: z.string() }))(({ id }) => ({
       resultA: id,
     }))
-    const b = df.make(z.object({ id: z.string() }))(({ id }) => ({
+    const b = withSchema(z.object({ id: z.string() }))(({ id }) => ({
       resultB: id,
     }))
 
@@ -148,10 +148,10 @@ describe('merge', () => {
   })
 
   it('should combine the error messages when both functions fail', async () => {
-    const a = df.make(z.object({ id: z.number() }))(() => {
+    const a = withSchema(z.object({ id: z.number() }))(() => {
       throw new Error('Error A')
     })
-    const b = df.make(z.object({ id: z.number() }))(() => {
+    const b = withSchema(z.object({ id: z.number() }))(() => {
       throw new Error('Error B')
     })
