+++ src/tests/trace.test.ts
@@ -5,12 +5,18 @@ import {
   it,
   z,
 } from '../test-prelude.ts'
-import { composable, df, trace, fromSuccess, success } from '../index.ts'
+import {
+  composable,
+  withSchema,
+  trace,
+  fromSuccess,
+  success,
+} from '../index.ts'
 import type { Composable } from '../index.ts'
 
 describe('trace', () => {
   it('converts trace exceptions to df failures', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => id + 1)
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => id + 1)
 
     const c = trace(() => {
       throw new Error('Problem in tracing')
@@ -43,7 +49,7 @@ describe('trace', () => {
   })
 
   it('intercepts inputs and outputs of a given domain function', async () => {
-    const a = df.make(z.object({ id: z.number() }))(({ id }) => id + 1)
+    const a = withSchema(z.object({ id: z.number() }))(({ id }) => id + 1)
 
     let contextFromFunctionA: unknown[] = []
